Rename private reflection helpers HasBit to IsPresent.
First of all, the HasBit(), SetBit(), ClearBit(), and SwapBit() methods are all
private so they are easier to change than a public-facing API.

The function named HasBit is called for both proto2 and proto3 fields. For
proto2 fields, it returns true if and only if the hasbit is set -- i.e. if the
field is present. For proto3 fields, it returns true if and only if the field
is nonempty -- i.e. if the field is present.

IsPresent is a more appropriate name for what this function is trying to do.
(I considered `HasPresence` too but that also leaves room for ambiguity: "has
presence" might be interpreted to refer to whether a field is an explicit
presence field or an implicit presence field).

See documentation here for what "field presence" means:
https://github.com/protocolbuffers/protobuf/blob/main/docs/field_presence.md

Note that in the proto3 case, what this function is doing has nothing to do
with bits!

I propose renaming to IsPresent(), SetPresence(), ClearPresence(), and
SwapPresence(), which should hopefully be less confusing and would describe the
fact that these methods work for implicit-presence fields.

PiperOrigin-RevId: 669055652